Consultancy/ seperating dataSource from viewController
