+++ TaskMVP/Scenes/Consultancy/CategoriesPresenter/ConsultancyView.swift
@@ -5,9 +5,10 @@
 //  Created by Ahmad Abdulrady on 29/07/2021.
 //
 
-import Foundation
+import UIKit
 
 protocol ConsultancyView: AnyObject {
+    var categoriesDataSource: UICollectionViewDataSource? {get set}
     func categoriesLoaded()
     func networkError(error: Error)
     func showLoadingIndicator()
